Dont disable the confirm buttons
Arbitrators reported that there are some cases where the confirm
payment or receipt buttons are disabled and they cannot confirm it.
Another issue is that sometimes the confirm msg does not arrive and
after button click its disabled so the message cannot be resent but
clicking the button again.
The software should handle such case and there is some automatic resend
but it seems it does not cover all cases. By keeping the button active
the user can click again and there is higher chance that such cases
get resolved by itself.
I am aware that this is not a good solution but to find out what
exactly caused the issue and to build a more resilient resend
functionality requires much more work and this seems to be a easy fix
which might avoid many disputes.