Remove check if trade is disputed
It can be that one user opened a dispute but in the meantime the other
peer sent or received the funds. Atm after a dispute has been opened the
user cannot confirm anymore. This restriction forces all cases to be
resolved in arbitration. If we relax that restriction to allow users
to still confirm after they opened a dispute we might reduce work load
for arbitrators as the payout can be done by the users if both have
confirmed. The arbitrator still need to close the case to that the
open support cases get closed, but the payout tx was already created
 by the users.