Fix offer editing for low volume offers
For very low volume offers, it can happen that the minimum buyer security
deposit by coin value is bigger than the maximum security deposit by
percentage of trade amount.
This caused the security deposit validity check in the offer editor to fail,
making it impossible to edit these offers.
This commit fixes the problem by setting the default percentage value to
the model instead of the calculated one in this specific case.