Change text field for median to average 30
The average 30 days price is more relevant information than the small
different between average and median price. To add a second tuple of
both would be another option would overload the screen with not that
relevant information.

In a follow up commit we want to add an ip/down indication so see
price trend.

Fixes also handling of 0 values.