Add bsq price trend indicator (#3226)
* Change text field for median to average 30

The average 30 days price is more relevant information than the small
different between average and median price. To add a second tuple of
both would be another option would overload the screen with not that
relevant information.

In a follow up commit we want to add an ip/down indication so see
price trend.

Fixes also handling of 0 values.

* Apply reformat (no code change)

* Add support for columnIndex

* Add price trend indicator

If average 30 days price is > av. 90 days price we show a green
arrow up, otherwise a red arrow down. If same we don't show an arrow.

* Fix dev test change

Forgot to change back a small change to test directions...

* Fix padding issues