GetBlocksReq is no CapRequiringPayload anymore
GetBlocksRequest is no CapabilityRequiringPayloadi anymore, because if it is, our DAO sync at startup is not going to work.
That is because when a fresh connection to a seed node is created, the connection does not know
- that the target is a seed node
- the capabilities of the target node.
Hence, a CapabilityRequiring GetBlocksRequest is therefore not sent to the freshly connected peer.

This fix is not final, however, this issue is not easily resolved. On the other hand, we have to get rid of the DAO sync issues at startup.