Add more checks for invalid trades and disputes
Add check that refund agent cannot get a case with a unconfirmed
deposit tx (refund agent need to check anyway those txs before doing
the payout but with that we add more protection).
Also add a check that the deposit tx is not null. We remove
the code for finding any multisig tx in case no deposit tx is available.
That caused anyway more troubles in disputes as the displayed deposit tx
could be from any other trade. Users cannot open mediation or arbitration
in such cases but as they cannot help anyway with the payout that seems
to be ok. users should use the keybase support channel.

If we have a unconfirmed deposit tx in closed trades we show a popup and
move it to failed trades.