Remove HashmapChangedListener::onBatch operations
Now that the only user of this interface has been removed, go ahead
and delete it. This is a partial revert of
f5d75c4f6085e17f512a73aff3f6d16f6dad5cc3 that includes the code that was
added into ProposalService that subscribed to the P2PDataStore.